just when you thought you were done...
added 2 setup pages to HTML - one for the car setup file, one for the device setup file. No need to open the case to pull the SD card, or leave an opening for USB access to the board to update these files.

This also required implementing LittleFS to put the setup and results HTML files in LittleFS for direct access by the web page server.

changed malloced pointers for settings tire and position info to fixed arrays - just easier at the expense of a little more memory use

Ready for Beta test of new hardware/software!